Refactor: 사이드바, 채팅 컴포넌트 정리
- 사이드바, 메인사이드바, 룸사이드바 정리
- 채팅->채팅카드, 채팅리스트로 채팅 정리
