+++ client/src/components/chat/ChatList/index.tsx
@@ -1,14 +1,13 @@
 import React, { useCallback, useEffect, useRef } from 'react';
-import styled from 'styled-components';
 import { TiDelete } from 'react-icons/ti';
+import socket from '@socket/socket';
+import { SocketEvents } from '@socket/socketEvents';
+import { Container, List, TextAreaWrapper, TextArea, Line, TextResetBtn } from './style';
+import ChatCard from '@components/chat/ChatCard';
 import { useUser } from '@contexts/userContext';
 import { useTheme } from '@contexts/themeContext';
 import useInput from '@hooks/useInput';
-import socket from '@socket/socket';
-import Chat from './Chat';
-import { INPUT, RoomType } from '@utils/constant';
-import { CHAT } from '@utils/styleConstant';
-import { SocketEvents } from '@socket/socketEvents';
+import { INPUT, RoomType, PLACEHOLDER_TXT, KEY_PRESS } from '@utils/constant';
 
 interface ChatListProps<T> {
   chats: Array<Record<string, T | undefined>>;
@@ -17,60 +16,7 @@ interface ChatListProps<T> {
   setChats: React.Dispatch<React.SetStateAction<Array<Record<string, T | undefined>>>>;
 }
 
-const Container = styled.div`
-  width: 100%;
-  height: 100%;
-  ${({ theme }) => theme.flexColumn};
-`;
-
-const List = styled.ul`
-  width: 100%;
-  height: ${CHAT.listHeight};
-  ${({ theme }) => theme.flexColumn};
-  padding: ${({ theme }) => theme.paddings.sm};
-  overflow: auto;
-  white-space: pre-wrap;
-`;
-
-const TextAreaWrapper = styled.div`
-  position: relative;
-  width: 100%;
-  height: ${CHAT.inputHeight};
-  ${({ theme }) => theme.flexColumn};
-  justify-content: end;
-  align-items: center;
-`;
-
-const TextArea = styled.textarea`
-  width: ${CHAT.inputWidth};
-  height: 6rem;
-  font-size: ${CHAT.fontSize};
-  padding: ${({ theme }) => theme.paddings.sm};
-  border: 2px solid ${({ theme }) => theme.colors.borderGrey};
-  border-radius: ${({ theme }) => theme.borderRadius.base};
-  background-color: transparent;
-  ::placeholder {
-    font-size: ${CHAT.fontSize};
-  }
-`;
-
-const Line = styled.div`
-  width: ${CHAT.inputWidth};
-  border-top: 1px solid ${({ theme }) => theme.colors.black};
-  opacity: 0.4;
-  margin: 0 auto;
-`;
-
-const InitBtn = styled.span`
-  position: absolute;
-  bottom: 1rem;
-  right: 3rem;
-  .icon:hover {
-    opacity: 0.9;
-  }
-`;
-
-const InitBtnStyle = {
+const TextResetBtnStyle = {
   fill: 'grey',
   opacity: 0.7,
   fontSize: '2.2rem',
@@ -92,7 +38,7 @@ const ChatList = ({ uuid, chats, setChats, roomType }: ChatListProps<string>): J
 
   const onKeyPress = useCallback(
     (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
-      if (e.key === 'Enter') {
+      if (e.key === KEY_PRESS.enter) {
         if (!e.shiftKey) {
           e.preventDefault();
           sendMessage();
@@ -121,7 +67,7 @@ const ChatList = ({ uuid, chats, setChats, roomType }: ChatListProps<string>): J
       <List ref={scrollRef}>
         {chats.length > 0 &&
           Object.values(chats).map((chat, idx) => (
-            <Chat
+            <ChatCard
               key={idx}
               chat={chat}
               bgChatBox={roomType === RoomType.campfire ? theme.colors.bgChatBox : undefined}
@@ -131,16 +77,16 @@ const ChatList = ({ uuid, chats, setChats, roomType }: ChatListProps<string>): J
       <Line />
       <TextAreaWrapper>
         <TextArea
-          placeholder="Message..."
+          placeholder={PLACEHOLDER_TXT.chat}
           value={message}
           onChange={onChangeMessage}
           onKeyPress={onKeyPress}
           maxLength={INPUT.chatMaxLen}
         />
         {message && (
-          <InitBtn>
-            <TiDelete style={InitBtnStyle} onClick={onResetMessage} />
-          </InitBtn>
+          <TextResetBtn>
+            <TiDelete style={TextResetBtnStyle} onClick={onResetMessage} />
+          </TextResetBtn>
         )}
       </TextAreaWrapper>
     </Container>
