+++ client/src/components/sideBar/Room/style.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const Tabs = styled.div`
+  display: flex;
+`;
